README.md
An ambitious and career-focused individual, I am highly interested in establishing my professional path through an entry-level position in the dynamic field of cybersecurity. With a strong desire not only to launch my own career, but also to contribute significantly to the overarching goals of the company, I am confident that I can bring a fresh perspective to cybersecurity challenges.

Having recently concluded an intensive Cybersecurity Boot Camp at the University of Birmingham, I have acquired a comprehensive understanding of cybersecurity concepts, tools, and best practices. This hands-on training, combined with my natural enthusiasm for technology, equips me with the necessary skills to tackle various cybersecurity issues effectively. I am eager to apply this knowledge in a practical setting and look forward to the possibility of contributing to your team.